// __generated__/sendMessage.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: sendMessage
// ====================================================

export interface sendMessage_sendMessage {
  __typename: "MutationResponse";
  ok: boolean;
  id: number | null;
}

export interface sendMessage {
  sendMessage: sendMessage_sendMessage | null;
}

export interface sendMessageVariables {
  payload: string;
  roomId?: number | null;
  userId?: number | null;
}